feat(core): add OpenWrt router build profile and matrix (#135)
* feat(core): add OpenWrt router build profile and matrix

- split optional build features for bundled SQLite, CoAP, and Tokio multi-thread runtime
- use AtomicUsize counters for 32-bit target compatibility
- gate CoAP-only code and multi-thread-only tests when features are disabled
- document embedded OpenWrt/router builds in README
- add OpenWrt SDK Docker matrix helper for six common targets

* fix(core): address OpenWrt build review feedback

Keep listen event ids u64-monotonic on 32-bit targets by moving next_event behind a small mutex while leaving diagnostic request ids on native AtomicUsize. This preserves Last-Event-ID replay semantics on router builds.\n\nTighten the OpenWrt matrix script: remove the x86 row continuation bug, avoid eval-based exports, isolate Cargo output under target/openwrt-matrix instead of core/target, and remove the unused rut241 feature marker.

* fix(core): tighten router review follow-ups

Use AtomicU64 for listen ids on targets with native 64-bit atomics and keep the mutex fallback only for 32-bit router targets. Document that CoAP is both compile-time and runtime opt-in. Chown OpenWrt matrix Cargo artifacts back to the host user after root-in-container builds.

* perf(core): inline listen event counter dispatch

Keep the default 64-bit listen notify path equivalent to the old AtomicU64 fetch_add path. The 32-bit mutex fallback remains cfg-only for targets without native 64-bit atomics.

+++ core/src/middleware.rs
@@ -33,7 +33,7 @@ pub(crate) async fn add_core_response_headers(
     mut req: axum::http::Request<Body>,
     next: Next,
 ) -> Response {
-    let request_id = core.next_request.fetch_add(1, Ordering::Relaxed) + 1;
+    let request_id = (core.next_request.fetch_add(1, Ordering::Relaxed) + 1) as u64;
     // Stash on request extensions so the FSM driver
     // (`pipeline::run`) reads the SAME id we'll stamp on the
     // response. Without this, the middleware and `pipeline::run`
